+++ src/configs/notifications.ts
@@ -1,12 +1,7 @@
 import OneSignal from "react-native-onesignal";
-import { navigate } from "../routes/rootNavigation";
-
-interface IConfigureNotifications {
-  signed: boolean;
-}
 
 const configureNotificationsHandlers = async (signed: boolean) => {
   OneSignal.setNotificationOpenedHandler(({ action, notification }) => {});
 };
 
-export { configureNotificationsHandlers };
+export { configureNotificationsHandlers, OneSignal };